moving renderers to static folder
