+++ static/ui/password.html
@@ -50,7 +50,7 @@
 
   
   <script>
-       require('../../src/client/password-renderer.js')
+       require('./password-renderer.js')
   </script>
 
 </body>
