+++ static/ui/volume.html
@@ -29,7 +29,7 @@
         </div>
     </form>
     <script>
-        require('../../src/client/volume-renderer.js')
+        require('./volume-renderer.js')
     </script>
 </body>
